+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 - `py_gql.builders` has been moved to `py_gql.sdl` and `build_schema_ignoring_extensions` has been removed.
 - `Tracer` has been replaced by the more general concept of `Instrumentation` which is now backing `ApolloTracing`.
+- `Directive` is not a subclass of `GraphQLType` anymore.
 
 ### Updated / Added